Added some comments to the files

+++ ReceiptGenerator/item.cpp
@@ -1,8 +1,12 @@
+//class that holds all the information of a particular item in inventory
+
 #include "item.h"
 #include <iostream>
+#include <string>
 
 using namespace std;
 
+//base constructor for item where no variables are given
 item::item()
 {
 	SKU = 0;
@@ -11,7 +15,8 @@ item::item()
 	name = "NULL";
 }
 
-item::item(int givenSKU, string givenName, int givenAmount, float givenPrice)
+//overloaded constructor for an item where everything is given
+item::item(int givenSKU, string givenName, int givenAmount, double givenPrice)
 {
 	SKU = givenSKU;
 	name = givenName;
@@ -20,45 +25,51 @@ item::item(int givenSKU, string givenName, int givenAmount, float givenPrice)
 }
 
 
-
+//give the item a new SKU
 void item::setSKU(int newSKU)
 {
 	SKU = newSKU;
 }
 
+//give the item a new name
 void item::setName(string newName)
 {
 	name = newName;
 }
 
 
-
+//sets new amount of item
 void item::setAmount(int newAmount)
 {
 	amount = newAmount;
 }
 
-void item::setPrice(float newPrice)
+//sets new price of item
+void item::setPrice(double newPrice)
 {
 	price = newPrice;
 }
 
+//returns the SKU of the item
 int item::getSKU()
 {
 	return SKU;
 }
 
+//returns the name of the item
 string item::getName()
 {
 	return name;
 }
 
+//returns the amount of the item on hand
 int item::getAmount()
 {
 	return amount;
 }
 
-float item::getPrice()
+//returns the price of the item
+double item::getPrice()
 {
 	return price;
 }
